Extract margin calculations to new module
